+++ api/src/main/java/org/nuc/course/controller/ArtController.java
@@ -29,7 +29,7 @@ public class ArtController {
     }
 
     @PostMapping("/delete")
-    public Result delete(@RequestParam Integer id) {
+    public Result delete(@RequestParam Long id) {
         artService.deleteById(id);
         return ResultGenerator.genSuccessResult();
     }
@@ -41,7 +41,7 @@ public class ArtController {
     }
 
     @PostMapping("/detail")
-    public Result detail(@RequestParam Integer id) {
+    public Result detail(@RequestParam Long id) {
         Art art = artService.findById(id);
         return ResultGenerator.genSuccessResult(art);
     }
