+++ api/src/main/java/org/nuc/course/controller/CommentController.java
@@ -29,7 +29,7 @@ public class CommentController {
     }
 
     @PostMapping("/delete")
-    public Result delete(@RequestParam Integer id) {
+    public Result delete(@RequestParam Long id) {
         commentService.deleteById(id);
         return ResultGenerator.genSuccessResult();
     }
@@ -41,7 +41,7 @@ public class CommentController {
     }
 
     @PostMapping("/detail")
-    public Result detail(@RequestParam Integer id) {
+    public Result detail(@RequestParam Long id) {
         Comment comment = commentService.findById(id);
         return ResultGenerator.genSuccessResult(comment);
     }
