+++ api/src/main/java/org/nuc/course/controller/SizeController.java
@@ -29,7 +29,7 @@ public class SizeController {
     }
 
     @PostMapping("/delete")
-    public Result delete(@RequestParam Integer id) {
+    public Result delete(@RequestParam Long id) {
         sizeService.deleteById(id);
         return ResultGenerator.genSuccessResult();
     }
@@ -41,7 +41,7 @@ public class SizeController {
     }
 
     @PostMapping("/detail")
-    public Result detail(@RequestParam Integer id) {
+    public Result detail(@RequestParam Long id) {
         Size size = sizeService.findById(id);
         return ResultGenerator.genSuccessResult(size);
     }
