+++ api/src/main/java/org/nuc/course/controller/TagController.java
@@ -29,7 +29,7 @@ public class TagController {
     }
 
     @PostMapping("/delete")
-    public Result delete(@RequestParam Integer id) {
+    public Result delete(@RequestParam Long id) {
         tagService.deleteById(id);
         return ResultGenerator.genSuccessResult();
     }
@@ -41,7 +41,7 @@ public class TagController {
     }
 
     @PostMapping("/detail")
-    public Result detail(@RequestParam Integer id) {
+    public Result detail(@RequestParam Long id) {
         Tag tag = tagService.findById(id);
         return ResultGenerator.genSuccessResult(tag);
     }
